# Conflicts:
#	DataAccessLayer/Migrations/ArtShareContextModelSnapshot.cs
#	Presentation/appsettings.json

+++ DataAccessLayer/Repository/LikeRepository.cs
@@ -1,6 +1,8 @@
 using DataAccessLayer.BussinessObject.IRepository;
+using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using ModelLayer.BussinessObject;
+using ModelLayer.DTOS.Request.Like;
 
 namespace DataAccessLayer.BussinessObject.Repository;
 
@@ -28,10 +30,20 @@ public class LikeRepository : ILikeRepository
         return await _context.Likes.Where(c => c.ArtworkId.Equals(id)).ToListAsync();
     }
 
-    public async Task AddLikeAsync(Like like)
+    public async Task<IActionResult> AddLikeAsync(LikeCreation likeCreation)
     {
-        _context.Likes.Add(like);
+        var likeToAdd = new Like
+        {
+            Id = Guid.NewGuid(),
+            AccountId = likeCreation.AccountId,
+            ArtworkId = likeCreation.ArtworkId,
+            Artwork = _context.Artworks.FirstOrDefault(a => a.Id == likeCreation.ArtworkId),
+            CreateDate = DateTime.Now
+        };
+        likeToAdd.Artwork.Likes += 1;
+        _context.Likes.Add(likeToAdd);
         await _context.SaveChangesAsync();
+        return new StatusCodeResult(201);
     }
 
     public async Task UpdateLikeAsync(Like like)
@@ -40,10 +52,45 @@ public class LikeRepository : ILikeRepository
         await _context.SaveChangesAsync();
     }
 
-    public async Task DeleteLikeAsync(Guid id)
+    public async Task DeleteLikeAsync(Guid id, Guid ArtworkId)
     {
         var like = await _context.Likes.FindAsync(id);
-        _context.Likes.Remove(like);
-        await _context.SaveChangesAsync();
+        if (like != null)
+        {
+            var artwork = await _context.Artworks.FindAsync(ArtworkId);
+
+            if (artwork != null)
+            {
+                artwork.Likes -= 1;
+                _context.Likes.Remove(like);
+                await _context.SaveChangesAsync();
+            }
+        }
+    }
+
+    public async Task<Guid?> GetLikeId(Guid accountId, Guid artworkId)
+    {
+        // Tìm like có accountId và artworkId t??ng ?ng trong c? s? d? li?u
+        var like = await _context.Likes.FirstOrDefaultAsync(l => l.AccountId == accountId && l.ArtworkId == artworkId);
+
+        // N?u tìm th?y like, tr? v? Id c?a like, ng??c l?i tr? v? null
+        if (like != null)
+        {
+            return like.Id;
+        }
+        else
+        {
+            return null;
+        }
+    }
+
+
+    public async Task<bool> CheckIfLikeExists(Guid accountId, Guid artworkId)
+    {
+        // Ki?m tra xem có b?n ghi nào trong c? s? d? li?u có tài kho?n và ID tác ph?m t??ng ?ng không
+        var like = await _context.Likes.FirstOrDefaultAsync(l => l.AccountId == accountId && l.ArtworkId == artworkId);
+
+        // N?u có, tr? v? true, ng??c l?i tr? v? false
+        return like != null;
     }
 }
